// src/app/components/models/grupos.interface.ts
export interface Grupo {
    idGrupo: number;//Estos campos deben llamarse igual que en la clase del backend
    idCarrera: number;
    idMateria: number;//Estos campos deben llamarse igual que en la clase del backend
    idProfesor: number;
    ciclo: number;//Estos campos deben llamarse igual que en la clase del backend
    anio: number;
}